Added near_block.rs based on crates.io types
