+++ refiner-types-crates-io/Cargo.toml
@@ -10,6 +10,8 @@ license.workspace = true
 publish = false
 
 [dependencies]
+borsh.workspace = true
+near-lake-framework.workspace = true
 near-crypto = "0.30.1"
 near-indexer-primitives = "0.30.1"
 near-primitives = "0.30.1"
