+++ refiner-types-crates-io/src/lib.rs
@@ -1,17 +1,47 @@
-// Re-export of those types from crates.io that are private when accessed from near_lake_framework::*
-// to make them available when converting to near_primitives::*
-// See refiner-types/src/near_block.rs for more details
-pub use near_crypto::PublicKey as PublicKeyCratesIo;
-pub use near_crypto::Secp256K1Signature as Secp256K1SignatureCratesIo;
-pub use near_crypto::Signature as SignatureCratesIo;
+use near_indexer_primitives::StreamerMessage;
 
-pub use near_primitives::action::GlobalContractIdentifier as GlobalContractIdentifierCratesIo;
-pub use near_primitives::types::ShardId as ShardIdCratesIo;
-pub use near_primitives::errors::TxExecutionError as TxExecutionErrorCratesIo;
+pub mod near_block;
+pub use near_block::NEARBlock;
 
-/// NOTE: temporary added here
-///  Convert between types that have the same json representation
-pub fn ch_json<U: serde::Serialize, V: serde::de::DeserializeOwned>(input: U) -> V {
-    let value = serde_json::to_value(input).unwrap();
-    serde_json::from_value(value).unwrap()
+pub fn convert(message: StreamerMessage) -> NEARBlock {
+    NEARBlock {
+        block: message.block.into(),
+        shards: message.shards.into_iter()
+            .map(|indexer_shard| {
+                let chunk = indexer_shard.chunk.map(|chunk| {
+                    near_block::ChunkView {
+                        author: chunk.author,
+                        header: chunk.header.into(),
+                        transactions: chunk.transactions
+                            .into_iter()
+                            .map(|tx| {
+                                near_block::TransactionWithOutcome {
+                                    transaction: tx.transaction.into(),
+                                    outcome: near_block::ExecutionOutcomeWithOptionalReceipt {
+                                        execution_outcome: tx.outcome.execution_outcome.into(),
+                                        receipt: tx.outcome.receipt.map(Into::into),
+                                    },
+                                }
+                            })
+                            .collect(),
+                        receipts: chunk.receipts.into_iter().map(Into::into).collect(),
+                    }
+                });
+                near_block::Shard {
+                    shard_id: indexer_shard.shard_id,
+                    chunk,
+                    receipt_execution_outcomes: indexer_shard.receipt_execution_outcomes
+                        .into_iter()
+                        .map(|r| {
+                            near_block::ExecutionOutcomeWithReceipt {
+                                execution_outcome: r.execution_outcome.into(),
+                                receipt: r.receipt.into(),
+                            }
+                        })
+                        .collect(),
+                    state_changes: indexer_shard.state_changes,
+                }
+            })
+            .collect()
+    }
 }
